requirements.txt
fastapi
uvicorn[standard]
pydantic
pydantic-settings
openai
httpx
PyYAML
Jinja2
loguru
python-dotenv